+++ Sources/Behaviours/ApplicationEventBehaviour.swift
@@ -23,8 +23,8 @@ public final class ApplicationEventBehavior: ViewControllerLifeCycleBehaviour {
 		willEnterForeground = foreground
 		willEnterBackground = background
 
-		NotificationCenter.default.addObserver(self, selector: #selector(applicationWillEnterForeground(_:)), name: .UIApplicationWillEnterForeground, object: nil)
-		NotificationCenter.default.addObserver(self, selector: #selector(applicationWillResignActive(_:)), name: .UIApplicationWillResignActive, object: nil)
+		NotificationCenter.default.addObserver(self, selector: #selector(applicationWillEnterForeground(_:)), name: UIApplication.willEnterForegroundNotification, object: nil)
+		NotificationCenter.default.addObserver(self, selector: #selector(applicationWillResignActive(_:)), name: UIApplication.willResignActiveNotification, object: nil)
 	}
 
 	@objc