+++ Sources/Core/Application Services/ConfigureMainQueueApplicationService.swift
@@ -12,7 +12,7 @@ import UIKit
 /// Internal class for configuring the main queue for checking later
 final class ConfigureMainQueueApplicationService: NSObject, ApplicationService {
 	// swiftlint:disable:next discouraged_optional_collection
-	func application(_ application: UIApplication, willFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]? = nil) -> Bool {
+	func application(_ application: UIApplication, willFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]? = nil) -> Bool {
 		DispatchQueue.configureMainQueue()
 
 		return true